MitraAI-Server/requirements.txt
Flask
Flask-Cors
google-auth
google-auth-oauthlib
google-cloud
google-cloud-aiplatform
langchain
requests
langchain_google_vertexai
pdfplumber
python-dotenv
